Move express care window actions and reducers to appointments so we can do lazy loading properly
