+++ src/applications/vaos/api/request_eligibility_criteria.json
@@ -38,7 +38,7 @@
               },
               {
                 "day": "FRIDAY",
-                "canSchedule": false,
+                "canSchedule": true,
                 "startTime": "08:58",
                 "endTime": "23:59"
               },
