+++ src/applications/vaos/containers/AppointmentsPage.jsx
@@ -11,11 +11,9 @@ import {
   confirmCancelAppointment,
   startNewAppointmentFlow,
   fetchFutureAppointments,
-} from '../actions/appointments';
-import {
   fetchExpressCareWindows,
-  startNewExpressCareFlow,
-} from '../actions/expressCare';
+} from '../actions/appointments';
+import { startNewExpressCareFlow } from '../actions/expressCare';
 import CancelAppointmentModal from '../components/cancel/CancelAppointmentModal';
 import {
   getCancelInfo,
