+++ src/applications/vaos/containers/NewExpressCareRequestLayout.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router';
 import LoadingIndicator from '@department-of-veterans-affairs/formation-react/LoadingIndicator';
 import Breadcrumbs from '../components/Breadcrumbs';
 import NeedHelp from '../components/NeedHelp';
-import * as actions from '../actions/expressCare';
+import * as actions from '../actions/appointments';
 import { FETCH_STATUS } from '../utils/constants';
 import { selectExpressCare } from '../utils/selectors';
 import ErrorMessage from '../components/ErrorMessage';
