+++ src/applications/vaos/reducers/expressCare.js
@@ -6,9 +6,6 @@ import {
 import set from 'platform/utilities/data/set';
 
 import {
-  FETCH_EXPRESS_CARE_WINDOWS_FAILED,
-  FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED,
-  FETCH_EXPRESS_CARE_WINDOWS,
   FORM_ADDITIONAL_DETAILS_PAGE_OPENED,
   FORM_DATA_UPDATED,
   FORM_PAGE_CHANGE_COMPLETED,
@@ -22,16 +19,9 @@ import {
   FORM_SUBMIT,
 } from '../actions/expressCare';
 
-import {
-  FETCH_STATUS,
-  EXPRESS_CARE,
-  WEEKDAY_INDEXES,
-} from '../utils/constants';
-import { STARTED_NEW_EXPRESS_CARE_FLOW } from '../actions/sitewide';
+import { FETCH_STATUS } from '../utils/constants';
 
 const initialState = {
-  windowsStatus: FETCH_STATUS.notStarted,
-  supportedFacilities: null,
   newRequest: {
     data: {},
     pages: {},
@@ -115,48 +105,6 @@ export default function expressCareReducer(state = initialState, action) {
         },
       };
     }
-    case FETCH_EXPRESS_CARE_WINDOWS:
-      return {
-        ...state,
-        windowsStatus: FETCH_STATUS.loading,
-      };
-    case FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED: {
-      const { settings } = action;
-      // We're only parsing out facilities in here, since the rest
-      // of the logic is very dependent on the current time and we may want
-      // to re-check if EC is available without re-fecthing
-      const supportedFacilities = settings
-        // This grabs just the facilities where EC is supported
-        .filter(
-          facility =>
-            facility.customRequestSettings?.find(
-              setting => setting.id === EXPRESS_CARE,
-            )?.supported,
-        )
-        // This makes sure we only pull the days where EC is open
-        .map(facility => ({
-          facilityId: facility.id,
-          days: facility.customRequestSettings
-            .find(setting => setting.id === EXPRESS_CARE)
-            .schedulingDays.filter(day => day.canSchedule)
-            .map(daySchedule => ({
-              ...daySchedule,
-              dayOfWeekIndex: WEEKDAY_INDEXES[daySchedule.day],
-            }))
-            .sort((a, b) => (a.dayOfWeekIndex < b.dayOfWeekIndex ? -1 : 1)),
-        }));
-
-      return {
-        ...state,
-        windowsStatus: FETCH_STATUS.succeeded,
-        supportedFacilities,
-      };
-    }
-    case FETCH_EXPRESS_CARE_WINDOWS_FAILED:
-      return {
-        ...state,
-        windowsStatus: FETCH_STATUS.failed,
-      };
     case FORM_FETCH_REQUEST_LIMITS: {
       return {
         ...state,
@@ -243,13 +191,6 @@ export default function expressCareReducer(state = initialState, action) {
         submitStatus: FETCH_STATUS.failed,
         submitErrorReason: action.errorReason,
       };
-    case STARTED_NEW_EXPRESS_CARE_FLOW: {
-      return {
-        ...initialState,
-        windowsStatus: state.windowsStatus,
-        supportedFacilities: state.supportedFacilities,
-      };
-    }
     default:
       return state;
   }
