+++ src/applications/vaos/tests/utils/selectors.unit.spec.js
@@ -611,8 +611,8 @@ describe('VAOS selectors', () => {
         .add('1', 'minutes')
         .tz('America/Denver');
       const state = {
-        expressCare: {
-          supportedFacilities: [
+        appointments: {
+          expressCareFacilities: [
             {
               facilityId: '983',
               days: [
@@ -648,8 +648,8 @@ describe('VAOS selectors', () => {
         .subtract('1', 'minutes')
         .tz('America/Denver');
       const state = {
-        expressCare: {
-          supportedFacilities: [
+        appointments: {
+          expressCareFacilities: [
             {
               facilityId: '983',
               days: [
@@ -681,8 +681,8 @@ describe('VAOS selectors', () => {
         .add('2', 'minutes')
         .tz('America/Denver');
       const state = {
-        expressCare: {
-          supportedFacilities: [
+        appointments: {
+          expressCareFacilities: [
             {
               facilityId: '983',
               days: [
@@ -724,8 +724,8 @@ describe('VAOS selectors', () => {
         .subtract(1, 'minutes')
         .tz('America/Denver');
       const state = {
-        expressCare: {
-          supportedFacilities: [
+        appointments: {
+          expressCareFacilities: [
             {
               facilityId: '983',
               days: [
@@ -770,8 +770,8 @@ describe('VAOS selectors', () => {
         .add(-1, 'minutes')
         .tz('America/Denver');
       const state = {
-        expressCare: {
-          supportedFacilities: [
+        appointments: {
+          expressCareFacilities: [
             {
               facilityId: '983',
               days: [
